Massive hack/workaround.
When an insert is typed with a \\n in the middle, Selenium types the \\n
correctly, but everything following has a \\n prepended and the cursor stays in
plage. For example, inserting "a\\nbcd" yields "a\nd\nc\nb". This commit
addresses the problem by moving the cursor to the head of the document and then
back in place whenever we encounter a \\n. This is definitely treating the
symptom and not the problem, but I can't figure out what's causing this behavior
so putting this in place until I figure out something better.