Do not autoformat link attr when appending to a link.
Currently, if you append to hyperlinked text, Scribe doe not continue the
hyperlink automatically. I think this is a bug, and so once we acknowledge
that/make the fix, I'll take out this hack. But right now it's useful so that we
don't hit that error case all the time.