Get rid of mergeAdjacent optitmization
Requiring document to merge adjacent nodes to be correct is
expensive. It and removeRedundant needs to be run
together or continuously to guarantee correctness. Running together
introduces a lot of code complexity and running after one another is
computationally expensive. Document still works correctly without
optimization so not worth including it.