Another attempt to rework selection to fix bugs
Main change is we now store a selection object and when asked for one
and we do not have focus, we return the saved selection. We also no
longer set the selection to on update. This change may or may not have
been neccessary but follows the principle to not interfere with the
browser's native selection behavior as much as possible.