Fix #67
When a click is detected on a toolbar item we transfer focus back to the
editor. This focus() call seems to disable the editor in iOS. A simple
contenteditable div inside an iframe on jsfiddle exhibits the same
behavior indicating a browser bug. As a workaround we prevetDefault on
mousedown which prevents the transfer of focus. We still get the click
event so we know what was clicked.