Fix bug on inserting text on empty line in IE
Used to create another line. Reason is since we display none on the
break tag, it just shows the following break thus when we put the cursor
on the empty line IE thinks its after the break tag and thus when we
type its inserted after and once there's a printable character two lines
show. Solution is to get rid of break tags on normalization in IE even
if there is content after it.