Fix #139
IE11 uses <strong> tags so if a line ends with bolded text and cursor is
at the end and you hit enter the result is a new <p><strong></strong><p>
which selection position encoding cannot handle as the strong tag will
be removed. We must insert our own immutable node as a child to preserve
cursor location (which will be removed on line optimization).