Use event.clipboardData to implement onpaste
This seems to behave better in recent versions of Firefox, where the
native behavior happens before the .ql-paste-manager div can be focused.

All recent versions of major browsers now support accessing clipboardData
from the paste event, according to http://stackoverflow.com/questions/2176861.

However, some browsers (e.g. Safari on Mac) do not return anything in the call
to getData('text/html'). If this is the case, fall back to the previous logic
of changing focus and allowing the default behavior.