In applyDelta(), emit the change for localDelta before continuing
This fixes a bug where the changes represented by `localDelta` would be emitted
twice, because in that case the result of _trackDelta() will effectively be
`delta.compose(localDelta)`.

I thought it was clearest to use checkUpdate() to trigger the text-change event
for any localDelta before proceeding. In order to do this, I modified
`checkUpdate()` to return the delta found.

An alternative solution would be to continue using _update(), and simply let
there be a single text-change event emitted, for the composition of `localDelta`
and the `delta` parameter. I thought two separate events felt cleaner, however.

I've added some assertions to the local update test to demonstrate the expected
behavior, and also to verify the return value of `checkUpdate()`.