Always use native copy and cut
Removes an optimization that stores and applies a delta when able. Since
we do not know where the copied or cut content will be pasted we always
need to fill in text/plain and text/html. When the browser fills these
in they get rid of our application/json clipboard data.

Fixes #785