Force selection update on format
We do not update selection natively when the node and offset has not
changed, however issues appear to arise when its parentage has changed
in the case of formatting. In Chrome there is sometimes a cursor
rendering bug, in Firefox the next character will be in the incorrect
location.

Fix #803