fix line merge behavior
If we have <h1>0</h1><h2>2</h2> and delete to newline between the two
headers, should the resulting line be h1 or h2? What if they were
bullets vs lists, or other line formats? Testing on word processors seem
to depend on the format itself, rather than always keeping the first
line's format or the 2nd line's. This does not work well for a
customizable and consistent editor because we do not know the formats
ahead of time, so cannot order based on format. We must either always
take the first line's format or the second line's. Prior to this commit
the editor favored the first and the Delta format favored the second. We
standardize to always favor the second.

Fixes #889