Use KeyboardEvent.key instead of .keyCode
Changes the Keyboard module to make use of the KeyboardEvent.key
event property rather than the now obsolete KeyboardEvent.keyCode
property. This has several advantages and is discussed in issue
1091.

THIS CHANGE IS A BREAKING CHANGE.
no attempt has been made to maintain backwards compatibility.

Existing unit tests have been modified to reflect the new behaviour
and pass. The functional test required no modification to pass.
Perhaps some more tests should be written to express the new
possibilities afforded by this change. For example, a test that
makes use of listening for the '@' symbol keypress reliably.

Note: Maybe an 'ignoreCase' option should be used in
Keyboard.addBinding to allow listening for both 'a' and 'A'.
Unlike the keyCode property, the key property is case sensitive.