Ensuring Quill picker is fully accessible
Previously, the Quill picker was not a part of the tab order, which makes the picker inaccessible to screen readers/keyboard only users. Furthermore, once the picker was added to the tab order, I found that the picker options were not tabbable either, so I made the option items tabbable as well. I also made the dropdown escapable, so that screen reader/keyboard only users who opened the dropdown could exit out of it. Finally, I added the appropriate ARIA attributes to make sure screen readers could describe the state of the picker properly.