Add length to FormulaBlot
Currently it's impossible to select inside a formula because its
length is 1, so LinkedList.find can only find it when selecting right
before the formula. This change adds a length() method to FormulaBlot
which is equal to the text typed by the user, so selecting a formula
will work properly.