Don't select tooltip when editing formula
Tooltip.edit selects the textbox. This is fine for existing links,
since the user has to click "Edit" to edit the link, and for adding
formulas and links, but not for editing formulas: for example,
clicking right before the formula will open the tooltip, which is a
bit annoying and makes it impossible to add text right before the
formula.

This change adds a `create` option to Tooltip.edit. Only when it's
true will the textbox be selected.