Move drag blocking to `Scroll`
At the moment, consumers cannot enable drag events within the editor,
because they're all disabled by the `Quill` constructor registering an
anonymous `dragstart` handler.

This change allows consumers to override this behaviour by moving this
listener to the `Scroll` blot, where the `dragstart` is prevented in a
named method, which can be overridden by extending the class.