Fix list keyboard left/right arrow handling
At the moment, if I position my cursor at the start of a list item and
hit left, or at the end of a list item and hit right, the cursor does
not immediately move. This is because there is a UI element in the DOM
(the bullet/number).

This change adds a default keyboard handler for this case, which reuses
code from the similar embed handler. We check for the presence of a UI
element with the class defined on `ParentBlot`, making the assumption
that this element will appear at the [start of the blot][1].

[1]: https://github.com/quilljs/parchment/blob/4257051dfc0a9c3e240c4906f8e36a697f95b1f8/src/blot/abstract/parent.ts#L41