Make `Selection` into a module
Sometimes consumers of Quill want to modify the behaviour of the
`Selection` class. However, unlike all the other default behaviour of
Quill, such as `Keyboard`, `Clipboard` and `History`, `Selection` is not
a `Module`, and so its behaviour cannot be modified without forking or
monkey-patching.

This change moves the `Selection` class into the `modules` directory
alongside the other core Quill modules, so that its behaviour can be
overridden like any other module.