Fix insert inline embed with delete before block embed
This fix requires the delta to explicitly specify the implicit newline
that gets added as part of the new block container that the inline embed
is created inside. If not explicitly specified, the internal index
tracking will be off by one and the delete will consume the implict
newline instead.