Alternate way of handling text replacement.
Remove the need of mutation observer in case of selected text getting replaced on keypress. Making the behaviour exactly same as first deleting a range of text & then inserting a char.

Why is this required? - Because we are entirely dependent on the browser dom to make the changes correctly and if there are any mistakes, then we need to reactively make the changes required after looking at mutation records from the browser. This is specially a nightmare when we are dealing with custom blots/attributors. The proposed changes will bypass the need of mutation observer altogether!