🐛 Run clipboard matchers against plain text pastes
This is a reimplementation of https://github.com/quilljs/quill/pull/3530

At the moment, if the clipboard pastes `text/plain` content and no
`text/html` content, the `Clipboard.convert()` function will completely
skip the matching logic.

This is surprising when registering text node clipboard matchers.

This change updates the `convert()` function to match the plain text
against the plain text matchers, just like we do with HTML.

Note that these matchers will run _before_ applying the formats for
["paste and match style"][1], so they won't match an element matchers
for the target formatting (which I think should be expected anyway).

[1]: https://github.com/quilljs/quill/pull/3927